Update / move unit tests for path parameter extraction
